Added displayUserProfile() to js files
