+++ Projects.html
@@ -27,37 +27,8 @@
   </div>
 
   <!-- USER PROFILE -->
-  <div class="profile-card card" style="width: 18rem;">
-    <img class="card-img-top" src="..." alt="Card image cap">
-    <div class="card-body">
-      <h4 class="card-title">Name</h4>
-      <h5 class="card-text">username</h5>
-      <p>Bio</p>
-      <button type="button" class="btn btn-secondary">Follow</button>
-      <button type="button" class="btn btn-secondary">Sponsor</button>
-      <button type="button" class="btn btn-secondary">...</button>
-      <p>followers</p>
-      <p>following</p>
-    </div>
-    <ul class="list-group list-group-flush">
-      <li>New York, New York</li>
-      <li>github@aboutmonica.com</li>
-      <li>https://www.aboutmonica.com</li>
-      <li>@indigitalcolor</li>
-    </ul>
-    <ul class="list-group list-group-flush">
-      <h6>Highlights</h6>
-      <li>Arctic Code Vault Contributor</li>
-      <li>GitHub Star</li>
-      <li>PRO</li>
-    </ul>
-    <ul class="list-group list-group-flush">
-      <h6>Oragnizations</h6>
-    </ul>
-    <ul class="list-group list-group-flush">
-      <h6>Sponsors</h6>
-    </ul>
-  </div>
+  <div id="userProfile"></div>
+
 <div id="loopContainer" class="loopCard loopContainer"></div>
 <form id="projectFormContainer" class="formStyle projectForm" >
   <div class="card">
